refactor(analysis): extract function counting helpers
